pandemapDjango/requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
dj-static==0.0.6
Django==3.0.8
django-crontab==0.7.1
django-taggit==1.3.0
djangorestframework==3.11.0
docutils==0.15.2
gunicorn==19.1.1
ibm-cos-sdk==2.6.3
ibm-cos-sdk-core==2.6.3
ibm-cos-sdk-s3transfer==2.6.3
idna==2.10
jmespath==0.10.0
lomond==0.3.3
numpy==1.19.0
nltk==3.5
opencv-python==4.2.0.34
pandas==1.0.5
Pillow==7.2.0
psycopg2-binary==2.8.5
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.18
sqlparse==0.3.1
static3==0.7.0
tabulate==0.8.7
tqdm==4.48.0
urllib3==1.25.9
watson-machine-learning-client==1.0.378
